Renamed converter to be consistent with the log4j module
